feat(validate): add functionality to switch cwd to input file path dir, and reset after validations for relative validation paths. Add ChangeCwd and ValidateChecksum methods to common. Remove ComponentDefinitionDir from config in favor or ChangeCwd method.
